signedness fix
git-svn-id: svn://svn.code.sf.net/p/actiongame/code/trunk/ac@2986 e016b1c5-ed68-44aa-a2f7-31cac88cffde

+++ source/src/menus.cpp
@@ -205,7 +205,7 @@ struct mitemtextinput : mitemtext
         int tw = VIRTW/4;
         if(selection) renderbg(x+w-tw, y-FONTH/6, tw, NULL);
         draw_text(text, x, y);
-        draw_text(input.buf, x+w-tw, y, 255, 255, 255, 255, selection ? (input.pos>=0 ? input.pos : strlen(input.buf)) : -1);
+        draw_text(input.buf, x+w-tw, y, 255, 255, 255, 255, selection ? (input.pos>=0 ? input.pos : (int)strlen(input.buf)) : -1);
     }
 
     virtual void focus(bool on)
